feat: Employee, Department, EmployeeDepartment 매핑
- employee, department, employeeDepartment 매핑
- 기존 패스워드 대응BcryptPasswordEncoder 적용
- 기본 직원 리스트 조회 샘플

+++ src/main/java/com/example/proto/auth/JwtManager.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
 
 import java.util.Date;
-import java.util.Map;
 
 @Component
 public class JwtManager {
@@ -21,7 +20,7 @@ public class JwtManager {
 
     private static final String AUTH_HEADER = "Authorization";
 
-    public String createToken(Long empNo) {
+    public String createToken(Integer empNo) {
         return Jwts.builder()
             .setSubject(String.valueOf(empNo))
             .setIssuedAt(new Date())
